Rework the enqueueing and localizing of JavaScript
Instead of loop_end(), use wp_footer() for localizing trackserver.js,
because loop_end() can be called many times during a single request
under certain conditions.

Furthermore, use the_post() for a secondary way of detecting the
shortcode, for when the original detect_shortcode() algorithm fails.

This solves the case where a secondary WP_Query is in use, and the
main query has a different collection of posts than what is actually
shown.

This secondary detection should render the [tsscripts] shortcode redundant,
but the disadvantage of relying on it is, that the external CSS is loaded
at the bottom of the page, right before the scripts, instead of in <head>.

git-svn-id: http://plugins.svn.wordpress.org/trackserver/trunk@1314632 b8457f37-d9ea-0310-8a92-e5e31aec5664