Use the correct track ID for checking the nonce. Save mods for all tracks.
This may leave some kind of (really small) CSRF risk, because
the nonce that is sent, doesn't sign the track ID(s) that is/are being
modified.

Basically, the 'modifications' data can be freely changed in the POST
request while keeping the nonce and the value of 't' the same, and
the server will accept it.

This is because the request can contain modifications for any combination
of tracks, so it's not possibe to generate a nonce on the server before
it happens. Each track has its own nonce, and we use the first one that
/may/ (but does't have to) be in the set of modifications.

The risk is probably small, because the nonce has in fact been generated
by the WP admin and whatever track IDs are in the set of modifications,
we always check whether the current user has the right to edit them.
Bottom line is, the POST data /can/ be modified, but only by the user that
is using the WP admin and not by somebody on the outside.