Add bulk action to view (and edit) multiple tracks simultaneously
The implementation uses the old way of getting tracks, one HTTP request
per track. It would be cool to use the 'gettrack query' method for
getting all the tracks at once, but this would require some dirty
modifications to the nonce checking when handling the gettrack query.

Since the query would be constructed in the client, there is no way to
sign it with a nonce on the server.