Add new [tsprofile] function to allow the user's personal tracking url to be displayed on the the front end in the user's profile page.
Also support added for user=@@ in [tsmap]. This will display the tracks for the currently logged in user. This allows a template page to be created that can be used by all users. Without this, all users would have needed their own personal page so they could be 'the author'.